+++ hooks/update-advocates.ts
@@ -1,7 +1,7 @@
 import fs from 'fs'
 import util from 'util'
 
-import { fetchAdvocates } from './advocates-conversion-utils'
+import { fetchAdvocates } from './advocate-conversion-utils'
 
 export default async function (apiKey: any, outputFolder: string) {
   const advocates = await fetchAdvocates(apiKey)
